+++ ios/RNCalendarEvents.h
@@ -1,5 +1,15 @@
-#import <React/RCTBridgeModule.h>
 #import <React/RCTEventDispatcher.h>
 
+#ifdef RCT_NEW_ARCH_ENABLED
+
+#import "RNCalendarEventsSpec.h"
+@interface RNCalendarEvents : NSObject <NativeRNCalendarEventsSpec>
+@end
+
+#else
+
+#import <React/RCTBridgeModule.h>
 @interface RNCalendarEvents : NSObject <RCTBridgeModule>
 @end
+
+#endif